Make sure the buffer length used is encodable in int32

+++ src/ccid.h
@@ -29,7 +29,7 @@
 #include <stdint.h>
 
 uint32_t
-icc_compose(uint8_t *buf, uint32_t buffer_length, uint8_t msg_type, int32_t data_len, uint8_t slot, uint8_t seq,
+icc_compose(uint8_t *buf, uint32_t buffer_length, uint8_t msg_type, size_t data_len, uint8_t slot, uint8_t seq,
             uint16_t param, uint8_t *data);
 
 uint32_t
